+++ queries/exact_duplicates.sql
@@ -1,0 +1,2 @@
+SELECT * FROM duplicate_summary
+ORDER BY observation_count DESC, sha256;