+++ queries/no_date.sql
@@ -1,0 +1,3 @@
+SELECT * FROM media_with_provenance
+WHERE date_status = 'NO_DATE'
+ORDER BY current_relative_path;